Uplift class generator and scriptgen system writer to azuki-scriptgen
