+++ azuki-script-generation/src/main/kotlin/com/anaplan/engineering/azuki/script/generation/RunnableScenarioClassGenerator.kt
@@ -1,25 +1,36 @@
-package com.anaplan.engineering.azuki.tictactoe.adapter.scriptgen
+package com.anaplan.engineering.azuki.script.generation
 
+import com.anaplan.engineering.azuki.core.runner.RunnableScenario
+import kotlin.collections.plus
+import kotlin.collections.sorted
 import java.util.*
+import kotlin.reflect.KClass
 
-object TicTacToeRunnableScenarioClassGenerator {
+/**
+ * Generates full Kotlin JUnit runnable scenario classes given a scenario script.
+ */
+abstract class RunnableScenarioClassGenerator<S : RunnableScenario<*, *, *, *, *, *, *, *, *, *, *, *>>(
+    val adapterSpecificImports: List<String>,
+    val scenarioClass: KClass<S>,
+) {
 
     fun generate(
         className: String = "Generated_" + UUID.randomUUID().toString(),
         packageName: String = "",
         implementationVersions: Map<String, String> = emptyMap(),
-        scenarioScript: String,
+        scenarioScript: ScenarioScript,
     ) = RunnableScenarioClass(className, packageName, buildString {
         if (packageName.isNotEmpty()) appendLine("package $packageName").appendLine()
 
-        val imports = ticTacToeStandardImports + listOfNotNull("com.anaplan.engineering.azuki.core.runner.*",
+        val commonImports = listOfNotNull("com.anaplan.engineering.azuki.core.runner.*",
             "com.anaplan.engineering.azuki.core.system.*",
             "com.anaplan.engineering.azuki.core.scenario.Since".takeUnless { implementationVersions.isEmpty() },
-            "com.anaplan.engineering.azuki.tictactoe.dsl.TicTacToeRunnableScenario")
+            scenarioClass.qualifiedName)
+        val imports = commonImports + adapterSpecificImports
         imports.sorted().forEach { appendLine("import $it") }
         appendLine()
 
-        appendLine("class ${className.replace("-", "_")} : TicTacToeRunnableScenario() {")
+        appendLine("class ${className.replace("-", "_")} : ${scenarioClass.simpleName}() {")
         appendLine()
         appendLine("    @GeneratedScenario")
         if (implementationVersions.isNotEmpty()) {
@@ -28,11 +39,16 @@ object TicTacToeRunnableScenarioClassGenerator {
             appendLine("    )")
         }
         appendLine("    fun test() {")
-        appendLine(scenarioScript)
+        appendLine(scenarioScript.render {
+            indent = 2
+            inOuterBlock = true
+
+            // We're going to format the whole test-case anyway, so formatting twice is pointless
+            useFormatter = false
+        })
         appendLine("    }")
         appendLine("}")
     })
 }
 
 data class RunnableScenarioClass(val className: String, val packageName: String?, val definition: String)
-
