--- tictactoe/adapter-script-generation/src/main/kotlin/com/anaplan/engineering/azuki/tictactoe/adapter/scriptgen/TicTacToeStandardImports.kt
@@ -1,9 +1,0 @@
-package com.anaplan.engineering.azuki.tictactoe.adapter.scriptgen
-
-/**
- * Default imports that should be added to any tic-tac-toe script (generated or parsed).
- */
-val ticTacToeStandardImports = arrayOf(
-    "com.anaplan.engineering.azuki.tictactoe.dsl.*",
-    "com.anaplan.engineering.azuki.tictactoe.*"
-)
