+++ tictactoe/script-runner/src/main/kotlin/com/anaplan/engineering/azuki/tictactoe/scriptrunner/JUnitTestCaseWriter.kt
@@ -68,10 +68,7 @@ class JUnitTestCaseWriter(
         // TODO - add utility function for arbitrary name
         className = generatedTestClass ?: ("Generated_" + UUID.randomUUID().toString().replace("-", "_")),
         packageName = generatedTestPackage,
-        scenarioScript = TicTacToeScriptGeneration.scenario.verifiableFromOracle(baseScenario, answers).render {
-            indent = 2
-            useFormatter = false
-        },
+        scenarioScript = TicTacToeScriptGeneration.scenario.verifiableFromOracle(baseScenario, answers),
         implementationVersions = mapOf(
             testImplementation.implementationName to (testImplementation.version ?: "0.0.0"),
             verifyingImplementation.implementationName to (verifyingImplementation.version ?: "0.0.0"),
